working Store v1.0

+++ src/Store/auth-context.js
@@ -1,4 +1,6 @@
-import React, { useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
+
+let logoutTimer;
 
 const AuthContext = React.createContext({
   token: "",
@@ -8,25 +10,69 @@ const AuthContext = React.createContext({
   logout: () => {},
 });
 
+const calculateRemainingTime = (expirationTime) => {
+  const currentTime = new Date().getTime();
+  const adjExpirationTime = new Date(expirationTime).getTime();
+
+  const remainingDuration = adjExpirationTime - currentTime;
+
+  return remainingDuration;
+};
+
+const getStoredToken = () => {
+  const storedToken = localStorage.getItem("token");
+  const storedExpirationTime = localStorage.getItem("expirationTime");
+
+  const remainingTime = calculateRemainingTime(storedExpirationTime);
+
+  if (remainingTime <= 60000) {
+    localStorage.clear();
+    return null;
+  }
+
+  return {
+    token: storedToken,
+    remainingTime: remainingTime,
+  };
+};
+
 export const AuthContextProvider = (props) => {
-  const initialToken = localStorage.getItem("token");
+  const tokenData = getStoredToken();
+  let initialToken;
+  if (tokenData) {
+    initialToken = tokenData.token;
+  }
   const [token, setToken] = useState(initialToken);
   const [localId, setLocalId] = useState(null);
 
   const userIsLoggedIn = !!token;
 
-  const loginHandler = (token, localId) => {
+  const logoutHandler = useCallback(() => {
+    setToken(null);
+    setLocalId(null);
+    localStorage.clear();
+
+    if (logoutTimer) {
+      clearTimeout(logoutTimer);
+    }
+  },[]);
+
+  const loginHandler = (token, localId, expirationTime) => {
     setToken(token);
     setLocalId(localId);
     localStorage.setItem("token", token);
     localStorage.setItem("localId", localId);
+    localStorage.setItem("expirationTime", expirationTime);
+
+    const remainingTime = calculateRemainingTime(expirationTime);
+    logoutTimer = setTimeout(loginHandler, remainingTime);
   };
 
-  const logoutHandler = () => {
-    setToken(null);
-    setLocalId(null);
-    localStorage.clear();
-  };
+  useEffect(() => {
+    if (tokenData) {
+      logoutTimer = setTimeout(logoutHandler, tokenData.remainingTime);
+    }
+  }, [tokenData, logoutHandler]);
 
   const contextValue = {
     token: token,
